Fixed head and tail pointers for the 'queue' both now point to a spot in memory with the address of the next item [0:47] and the tuple that has been queued [63:48]

+++ BitSolver/ABitsC3.h
@@ -4,7 +4,7 @@
 #include "CSP_Bits.h"
 #include "qbit.h"
 
-int ABitsC3(uint *inQueue, CSP *csp);
+int ABitsC3(ulong *inQueue, CSP *csp);
 int revise(CSP *csp, ushort Xi, ushort Xj);
 
 #endif
